test the code of reading IMU and test the servo control

+++ ControlBoard/Hardware/tim/tim.h
@@ -3,7 +3,7 @@
 
 #include "main.h"
 
-void servoControl(TIM_HandleTypeDef *htim, uint8_t angle, uint8_t num);
+void servoControl(TIM_HandleTypeDef *htim, int angle, uint8_t num);
 void motorControl(TIM_HandleTypeDef *htim, int speed, uint8_t num);
 
 #endif
